requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ib-insync==0.9.86
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
asyncio-mqtt==0.16.1
python-dotenv==1.0.0
tzdata==2023.3
